[refactor] moved episodeImage related entities to episode domain
