+++ src/main/java/com/dopamingu/be/domain/episode/repository/EpisodeImageRepository.java
@@ -1,7 +1,7 @@
-package com.dopamingu.be.domain.image.dao;
+package com.dopamingu.be.domain.episode.repository;
 
 import com.dopamingu.be.domain.episode.domain.Episode;
-import com.dopamingu.be.domain.image.domain.EpisodeImage;
+import com.dopamingu.be.domain.episode.domain.EpisodeImage;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
